requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
aiogram
httpx
async-lru
jinja2